Implement publishable props in ComponentSet
