+++ src/components/component/Component.tsx
@@ -6,6 +6,7 @@ import {
     DefaultContainerProps,
     DefaultShapeProps,
     FrameSpecificProps,
+    PublishableProps,
     SelectionEventProps,
     StyleOf
 } from '../../types';
@@ -30,10 +31,6 @@ import { transformAutoLayoutToYoga } from '../../styleTransformers/transformAuto
 import { OnLayoutHandlerProps, useOnLayoutHandler } from '../../hooks/useOnLayoutHandler';
 import { useImageHash } from '../../hooks/useImageHash';
 
-interface DocumentationLink {
-    readonly uri: string;
-}
-
 export interface ComponentProps
     extends DefaultShapeProps,
         DefaultContainerProps,
@@ -42,7 +39,8 @@ export interface ComponentProps
         BorderProps,
         CornerProps,
         FrameSpecificProps,
-        OnLayoutHandlerProps {
+        OnLayoutHandlerProps,
+        PublishableProps {
     style?: StyleOf<
         GeometryStyleProperties &
             YogaStyleProperties &
@@ -51,8 +49,6 @@ export interface ComponentProps
             BorderStyleProperties
     >;
     nodeRef?: any;
-    description?: string;
-    documentationLinks?: ReadonlyArray<DocumentationLink>;
 }
 
 const Component: React.FC<ComponentProps> = props => {
