+++ src/mixins/publishableMixin.ts
@@ -1,0 +1,7 @@
+import { propsAssign } from '../helpers/propsAssign';
+import { PublishableProps } from '../types';
+
+export const publishableMixin = propsAssign<PublishableProps, PublishableProps>(['description', 'documentationLinks'], {
+    description: '',
+    documentationLinks: []
+});
