+++ src/renderers/componentset.ts
@@ -5,6 +5,7 @@ import { exportMixin } from '../mixins/exportMixin';
 import { blendMixin } from '../mixins/blendMixin';
 import { frameMixin } from '../mixins/frameMixin';
 import { sceneNodeMixin } from '../mixins/sceneNodeMixin';
+import { publishableMixin } from '../mixins/publishableMixin';
 
 const createNewComponentSet = () => {
     const component = figma.createComponent();
@@ -23,6 +24,7 @@ export const componentset = node => props => {
 
     frameMixin(componentSetNode)(props);
     sceneNodeMixin(componentSetNode)(props);
+    publishableMixin(componentSetNode)(props);
 
     return componentSetNode;
 };
